+++ src/main/java/com/umc/hwaroak/serviceImpl/MemberServiceImpl.java
@@ -1,21 +1,27 @@
-package com.umc.hwaroak.service;
+package com.umc.hwaroak.serviceImpl;
 
+import com.umc.hwaroak.authentication.MemberLoader;
 import com.umc.hwaroak.domain.Member;
-import com.umc.hwaroak.dto.MemberResponseDto;
+import com.umc.hwaroak.dto.response.MemberResponseDto;
 import com.umc.hwaroak.exception.GeneralException;
 import com.umc.hwaroak.repository.MemberRepository;
 import com.umc.hwaroak.response.ErrorCode;
+import com.umc.hwaroak.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 @Service
 @RequiredArgsConstructor
-public class MemberServiceImpl implements MemberService{
+public class MemberServiceImpl implements MemberService {
 
+    private final MemberLoader memberLoader;
     private final MemberRepository memberRepository;
 
     @Override
-    public MemberResponseDto.InfoDto getInfo(Long memberId) {
+    public MemberResponseDto.InfoDto getInfo() {
+
+        Long memberId = memberLoader.getCurrentMemberId();
+
         Member member = memberRepository.findById(memberId)
                 .orElseThrow(()->new GeneralException(ErrorCode.MEMBER_NOT_FOUND));
